math_question_generator/requirements.txt
python-docx>=0.8.11
matplotlib>=3.7.1